requirements.txt
click==7.1.2
joblib==1.1.0
python-dotenv>=0.5.1
scikit-learn==0.24.2
dataclasses==0.6
pyyaml==3.11
marshmallow-dataclass==8.3.0
pandas==1.1.5
boto3==1.17.49
mlflow==1.15.0
pytest==6.2.3
coverage==6.3.2
pytest-cov==3.0.0
flake8==4.0.1
black==20.8b0
fastapi==0.63.0
uvicorn==0.13.4
requests==2.27.1
pydantic==1.9.1
